requirements.txt
alembic==1.14.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.19.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
